Moves the keymaps and udates cmp
Previously many of my keymaps were handled by a setup file when
whichkeys is loaded. However, there was an issue with some of the
keymaps not loading fast enough. For example "<leader>e" did not load
quick enough for my liking.

Many of the keycodes which used to be handles by this setup file are now
handled in the keymaps.lua initialisation. Other keymaps have been moved
to the lsp initialiation function in lsp.lua (which in hindsight is
where they should have been from the beginning).

Some keymaps have been changed for cmp. In particular, <C-M-n> and
<C-M-p> now move down and up the documentation, respectively.