Fixes issues with cmp and lsp, and updates keymaps
Afeter updating to neovim 0.11.3, some parts of the config needed a bit
of attention to ensure compadability. In particular they needed to be
updated to use `vim.lsp.config`. Some changes have also been made to the
way that cmp is loaded. This will likely have to be chnaged in the
future as it currently uses deprecated functions.

This update also remaps the <Up> and <Down> keys so that they perform gk
and gj, respectively. This enables easier navigation. Maybe it would
drive other people crazy but this is my config: if you don't like it
then use your own config.